deleting booking

+++ database-docker/init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
-    user_id INTEGER PRIMARY KEY,
+    user_id BIGINT PRIMARY KEY,
     user_name VARCHAR(255) NOT NULL,
     roles TEXT DEFAULT '',
     subjects TEXT DEFAULT '',
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS users (
 
 CREATE TABLE IF NOT EXISTS bookings (
     booking_id INTEGER PRIMARY KEY,
-    user_id INTEGER,
+    user_id BIGINT,
     user_role TEXT NOT NULL,
     subjects TEXT NOT NULL,
     event_date VARCHAR(10),
